requirements.txt
streamlit
qdrant-client
sentence-transformers
libcst
shortuuid
GitPython